+++ spec/rspec/mocks/any_instance_spec.rb
@@ -444,6 +444,16 @@ module RSpec
           end
         end
 
+        it 'works with a BasicObject subclass' do
+          klass = Class.new(BasicObject) do
+            include ::Kernel # to make #method available, just like Delegator
+            def foo; end
+          end
+
+          klass.any_instance.should_receive(:foo)
+          klass.new.foo
+        end
+
         context "with argument matching" do
           before do
             klass.any_instance.should_receive(:foo).with(:param_one, :param_two).and_return(:result_one)