+++ Changelog.md
@@ -19,6 +19,8 @@ Bug fixes
 * Fix `missing_default_stub_error` message so array args are handled
   properly (Myron Marston).
 * Explicitly disallow `any_instance.unstub!` (Ryan Jones).
+* Fix `any_instance` stubbing so that it works with `Delegator`
+  subclasses (Myron Marston).
 
 ### 2.12.2 / 2013-01-27
 [full changelog](http://github.com/rspec/rspec-mocks/compare/v2.12.1...v.2.12.2)